+++ src/Dto/Actions/dummyFaceupActionDto.d.ts
@@ -1,0 +1,8 @@
+import { ActionDto } from './actionDto';
+
+import { PlayerPosition } from '../../Enum/playerPosition';
+
+export interface DummyFaceupActionDto extends ActionDto {
+    DummyPlayer: PlayerPosition;
+    Player: PlayerPosition;
+}