+++ src/Dto/CardGame/contractBridgeBidDto.d.ts
@@ -1,0 +1,6 @@
+import { BidDto } from './bidDto';
+
+export interface ContractBridgeBidDto extends BidDto
+{
+    Value: number;
+}