+++ src/Type/CardGameAnnounceSymbolModel.d.ts
@@ -1,9 +1,10 @@
-import { BidType } from '../Enum/bidType';
+import { BidTrump } from '../Enum/bidTrump';
 
 export interface CardGameAnnounceSymbolModel
 {
-    id: BidType;
+    id: BidTrump;
     key: string;
     tooltip: string;
     value: string;
+    htmlEntity: string;
 }